extract + refactor habit stats card, add to overview screen
